// source-solana-project/src/contexts/sorter/types.ts
export type SorterKey =
  | 'chart-token-selector'
  | 'gm-list'
  | 'dashboard-markets-list'
  | 'gm-token-selector'
  | 'leaderboard-accounts-table'
  | 'leaderboard-positions-table';

export type SorterConfig<
  // eslint-disable-next-line @typescript-eslint/no-redundant-type-constituents
  SortField extends string | 'unspecified' = 'unspecified',
> = {
  orderBy: SortField;
  direction: SortDirection;
  isDefault?: boolean;
};

export type SortDirection = 'asc' | 'desc' | 'unspecified';